+++ apps/common/models.py
@@ -3,8 +3,7 @@
 from django.contrib.gis.db import models
 from django.contrib.postgres import fields as pgmodels
 from apps.utils.time import format_delta
-from django.db.models.fields import DecimalField, DateField, DateTimeField
-from django.contrib.postgres.fields import ArrayField
+
 
 class Process(models.Model):
     """Process used to generate the result, e.g. measurement or
@@ -61,35 +60,6 @@ class Property(models.Model):
     class Meta:
         ordering = ['name']
         verbose_name_plural = "properties"
-
-    def __str__(self):
-        return self.name
-
-
-class TimeSeriesFeature(models.Model):
-    id_by_provider = models.CharField(
-        help_text="ID of the station used by provider.",
-        max_length=50,
-        editable=False
-    )
-
-    name = models.CharField(
-        help_text="Human-readable name of the station.",
-        max_length=50
-    )
-
-    geometry = models.PointField(
-        help_text="Spatial information about feature."
-    )
-
-    property_values = ArrayField(DecimalField(decimal_places=5, max_digits=15), null=True, blank=True)
-    property_anomaly_rates = ArrayField(DecimalField(decimal_places=5, max_digits=15), null=True, blank=True)
-    phenomenon_time_from = models.DateTimeField()
-    phenomenon_time_to = models.DateTimeField()
-    value_index_shift = models.IntegerField()
-
-    class Meta:
-        managed = False
 
     def __str__(self):
         return self.name
